Implemented delete file or text message

+++ server/ChatApp.WebAPI/Controllers/MessageController.cs
@@ -1,8 +1,12 @@
 ﻿using ChatApp.Core.Contracts;
 using ChatApp.Core.Models;
 using ChatApp.Core.Models.OutputDTOs;
+using ChatApp.Infrastructure.Data;
+using ChatApp.WebAPI.Hubs;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.SignalR;
+using MongoDB.Bson;
 
 namespace ChatApp.WebAPI.Controllers
 {
@@ -11,19 +15,34 @@ namespace ChatApp.WebAPI.Controllers
     [Authorize]
     public class MessageController : ControllerBase
     {
+        private readonly IHubContext<ChatHub> hubContext;
         private readonly IMessageService messageService;
         private readonly IUserRoomService userRoomService;
+        private readonly IRoomService roomService;
 
-        public MessageController(IMessageService messageService, IUserRoomService userRoomService)
+        public MessageController(IHubContext<ChatHub> hubContext, IMessageService messageService, IUserRoomService userRoomService,
+            IRoomService roomService)
         {
+            this.hubContext = hubContext;
             this.messageService = messageService;
             this.userRoomService = userRoomService;
+            this.roomService = roomService;
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteMessage(string id, [FromQuery] string userId)
+        public async Task<IActionResult> DeleteMessage(string id, [FromQuery] string userId, [FromQuery] string type)
         {
-            bool isUserAuthorizedToDeleteMessage = await messageService.IsUserMessageCreatorAsync(id, userId);
+            bool isUserAuthorizedToDeleteMessage = false;
+            if (type == "text")
+            {
+
+                isUserAuthorizedToDeleteMessage = await messageService.IsUserMessageCreatorAsync(id, userId);
+            }
+            else
+            {
+                RoomFileDTO file = await roomService.GetFileAsync(ObjectId.Parse(id));
+                isUserAuthorizedToDeleteMessage = file.SenderId == userId ? true : false;
+            }
 
             if (!isUserAuthorizedToDeleteMessage)
             {
@@ -32,7 +51,21 @@ namespace ChatApp.WebAPI.Controllers
 
             try
             {
-                await messageService.DeleteMessageAsync(id);
+                if (type == "text")
+                {
+                    Message message = await messageService.GetMessageAsync(id);
+                    await messageService.DeleteMessageAsync(id);
+                    await hubContext.Clients.Group(message.RoomId.ToString()).SendAsync("DeleteMessage", id, message.RoomId.ToString());
+                }
+
+                if (type == "file")
+                {
+                    RoomFileDTO file = await roomService.GetFileAsync(ObjectId.Parse(id));
+                    await roomService.DeleteFileAsync(ObjectId.Parse(id));
+                    await hubContext.Clients.Group(file.RoomId).SendAsync("DeleteMessage", id, file.RoomId.ToString());
+                }
+
+
             }
             catch (Exception ex)
             {
